# Improve performance for half-edge construction from `<:Connectivity` vector