--- a/src/topologies/halfedge.jl
+++ b/src/topologies/halfedge.jl
@@ -144,9 +144,9 @@ function HalfEdgeTopology(elems::AbstractVector{<:Connectivity}; sort=true)
 
   # sort elements to make sure that they
   # are traversed in adjacent-first order
-  elemsort = sort ? adjsort(elems) : elems
-  eleminds = sort ? indexin(elemsort, elems) : 1:length(elems)
-  adjelems::Vector{Vector{Int}} = map(collect ∘ indices, elemsort)
+  eleminds = sort ? adjsortperm(elems) : 1:length(elems)
+  adjelems::Vector{Vector{Int}} = map(collect ∘ indices ∘ Base.Fix1(getindex, elems),
+    eleminds)
 
   # start assuming that all elements are
   # oriented consistently (e.g. CCW)
@@ -332,42 +332,50 @@ Base.convert(::Type{HalfEdgeTopology}, t::Topology) = HalfEdgeTopology(collect(e
 # HELPER FUNCTIONS
 # -----------------
 
-function adjsort(elems::AbstractVector{<:Connectivity})
+function adjsortperm(elems::AbstractVector{<:Connectivity})
+  listi = collect(eachindex(elems)[2:end])
   # initialize list of adjacent elements
   # with first element from original list
-  list::Vector{Tuple{Vararg{Int}}} = map(indices, elems)
-  adjs = similar(list, 0)
-  push!(adjs, popfirst!(list))
+  adjs = Int[1]
+  sizehint!(adjs, length(elems))
 
-  # the loop will terminate if the mesh
-  # is manifold, and that is always true
-  # with half-edge topology
-  while !isempty(list)
+  # found minimizes adjacency discontinuities. if `found == true` for the last edge in an
+  # element, then we continue from that new element adjacent to that edge
+  found = false
+  while !isempty(listi)
     # lookup all elements that share at least
     # one vertex with the last adjacent element
-    found = false
-    vinds = last(adjs)
+    adj = last(adjs)
+    vinds::Tuple{Vararg{Int}} = indices(elems[adj])
     for i in vinds
       not_i = filter(!=(i), vinds)
-      for j in reverse(eachindex(list))
+      j = 1
+      while j ≤ length(listi)
         # equivalent to `length(vinds ∩ list[j]) > 1` but more efficient (no allocs(?))
-        any(==(i), list[j]) || continue
-        isdisjoint(not_i, list[j]) && continue
-
-        # implicitly `list[j]` contains `i` and at least one other vertex
-        found = true
-        push!(adjs, popat!(list, j))
+        elem = indices(elems[listi[j]])
+        if any(==(i), elem) && !isdisjoint(not_i, elem)::Bool
+          # `list[j]` contains `i` and at least one other vertex
+          push!(adjs, popat!(listi, j))
+          found = true
+          # don't increment j here because `popat!` just put the j+1 element at j
+          # (avoids the need to reverse the array)
+        else
+          j += 1
+          found = false
+          iter += 1
+        end
       end
     end
 
-    if !found && !isempty(list)
+    if !found && !isempty(listi)
       # we are done with this connected component
       # pop a new element from the original list
-      push!(adjs, popfirst!(list))
+      push!(adjs, popfirst!(listi))
+      found = false
     end
   end
 
-  map(connect, adjs)
+  adjs
 end
 
 function anyhalf(inds, half4pair)