--- a/src/topologies/halfedge.jl
+++ b/src/topologies/halfedge.jl
@@ -341,10 +341,13 @@ function adjsortperm(elems::AbstractVector{<:Connectivity})
   # remaining list of elements to process
   oinds = collect(2:length(elems))
 
+  # `found` minimizes adjacency discontinuities. if `found == true` for the last edge in an
+  # element, then we continue from that new element adjacent to that edge
+  found = false
+
   # lookup all elements that share at least
   # two vertices (i.e., edge) with the last
   # adjacent element
-  found = false
   while !isempty(oinds)
     lelem = elems[last(einds)]
     vinds = indices(lelem)
@@ -360,6 +363,8 @@ function adjsortperm(elems::AbstractVector{<:Connectivity})
         if any(==(v), vinds′) && !isdisjoint(v!, vinds′)
           found = true
           push!(einds, popat!(oinds, iter))
+          # don't increment j here because `popat!` just put the j+1 element at j
+          # (avoids the need to reverse the array)
         else
           found = false
           iter += 1