Using a ~6.6k vertex, ~13k Triangle  mesh, I measure the following speed and memory improvements with BenchmarkTools:

```julia
BenchmarkTools.Trial: 19 samples with 1 evaluation per sample.
 Range (min … max):  262.328 ms … 291.279 ms  ┊ GC (min … max): 3.70% … 8.65%
 Time  (median):     274.083 ms               ┊ GC (median):    3.97%
 Time  (mean ± σ):   272.500 ms ±   8.548 ms  ┊ GC (mean ± σ):  5.22% ± 2.65%

  ▃ █                      ▃                                     
  █▁█▁▇▇▁▁▁▁▁▁▁▇▁▁▁▁▁▁▁▁▁▇▇█▁▇▇▇▁▁▁▁▁▁▁▇▁▁▇▁▁▁▁▁▁▁▁▁▇▁▁▁▁▁▁▁▁▁▇ ▁
  262 ms           Histogram: frequency by time          291 ms <

 Memory estimate: 105.32 MiB, allocs estimate: 1971625.
```

```julia
BenchmarkTools.Trial: 25 samples with 1 evaluation per sample.
 Range (min … max):  171.102 ms … 265.974 ms  ┊ GC (min … max): 0.00% … 5.11%
 Time  (median):     183.287 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   207.502 ms ±  36.065 ms  ┊ GC (mean ± σ):  1.13% ± 2.26%

  ▁█   ▄                          ▁                 ▁            
  ██▆▆▆█▁▆▁▁▁▁▁▁▁▁▁▁▆▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▆▁▆▁▁▁▁▁▆█▆▁▁▁▆▁▆▁▁▆ ▁
  171 ms           Histogram: frequency by time          266 ms <

 Memory estimate: 19.98 MiB, allocs estimate: 150413.
```

Judgement:
```julia
BenchmarkTools.TrialJudgement: 
  time:   -23.85% => improvement (5.00% tolerance)
  memory: -81.03% => improvement (1.00% tolerance)
```