@juliohm 2025-04-08
That is very nice. Could you please share a MWE so that I can reproduce benchmark results and test the PR locally? :)