@halleysfifthinc 2025-04-08
[Quads and mixed faces meshes here](https://gist.github.com/halleysfifthinc/5b0d45a44f4dd451daca097aa35b9846).

```julia
using Meshes, GeoIO, BenchmarkTools

quads = GeoIO.load("quads.obj").geometry
quads_tris = GeoIO.load("quads_tris.obj").geometry

@benchmark topoconvert(HalfEdgeTopology, quads)
@benchmark topoconvert(HalfEdgeTopology, quads_tris)
```

`adjsort` is responsible for ~85% of the runtime and ~50% of allocations (count), but the `HalfEdgeTopology(::AbstractVector{Tuple{HalfEdge,HalfEdge}})` constructor contributes ~80% of total allocated amount (from the `Dict`s). So although [47fa3f4](https://github.com/JuliaGeometry/Meshes.jl/pull/1183/commits/47fa3f4d64001f4c4366dce9bcad90a534cbd2e2) only removes a few individual allocations, they are larger allocations and save a decent amount of memory.