@halleysfifthinc 2025-04-14
I reimplemented `adjsort` as `adjsortperm`, which now returns a sort permutation (faster, and removes the need for the `indexin` call, which was a non-negligible runtime contributor). Before/after benchmarks have been updated in the top comment (its quite a bit faster).

Along the way, I found a weird mesh/topo that is manifold, however, the old `adjsort` led to one face being missing from the final converted topology. Despite my best efforts, I could not determine what/where/why the problem was. [This gist](https://gist.github.com/halleysfifthinc/27236babe51bd01a204b093a70e33c8d) contains two meshes which are identical save for the triangle orders and orientations (same vertices). I am happy to include the bad mesh as a test if you wish.