@juliohm 2025-04-15
@halleysfifthinc thank you for refactoring the code.

I refactored it a bit more to make sure the new algorithm is easy to read. Do we still need to reintroduce type annotations to improve performance? Appreciate it if you could take another look. If type annotations are necessary, we could start with annotating the return type of `adjsortperm`, by typing `einds::Vector{Vector{Int}}` at the last line of the function.

If we could encapsulate these type annotations in the `adjsortperm` function, that would decrease our chances to break type instability in the future.