@juliohm 2025-04-15
> Along the way, I found a weird mesh/topo that is manifold, however, the old `adjsort` led to one face being missing from the final converted topology.

Do you mean that the old implementation had a bug? Is it present in the new implementation in this PR?