@halleysfifthinc 2025-04-17
Benchmarks updated (now specifically for `adjsort`/`adjsortperm`)!

Re: `connected_components` utility, I agree. I'd like to see it exported, or `public` at least, so I can use it and avoid the Graphs.jl dependency.

> Do you mean that the old implementation had a bug? Is it present in the new implementation in this PR?

Yes, the old `adjsort` was leading to one face not being added/recognized in the final `HalfEdgeTopology`. That specific problem is no longer present with the new `adjsortperm` implementation.

However, the old `adjsort` was only part of the problem, I believe the main `HalfEdgeTopology(::Vector{<:Connectivity}` function is incorrect/unreliable for certain mixes of face orientations. I have a mesh[^1] which has inconsistencies after being converted to a `HalfEdgeTopology` (`next` or `prev` edges not matching the current `elem`). I have been able to determine that the bug is related to inconsistent orientations among the faces, but I am struggling to find any logical errors in the existing orientation handling code here and above https://github.com/JuliaGeometry/Meshes.jl/blob/28c75552ef2bf3f8da9050a853f0a7025bed9217/src/topologies/halfedge.jl#L177-L182

[^1]:  I can't share the mesh, and I don't understand the cause of the bug well enough to simplify the mesh like I did for the meshes attached to [my earlier comment](https://github.com/JuliaGeometry/Meshes.jl/pull/1183#issuecomment-2803353936))