@halleysfifthinc 2025-04-29
**tl;dr**: The face ordering with the new `connected_components`/`adjsortperm` isn't adjacently sorted under the strictest definition (subsequent faces share at least one already existing edge). Obviously, it is your call whether this is acceptable, but the updated `HalfEdgeTopology(::Vector{<:Connectivity})` function handles the current (and previous) behavior. (Although nearly all of the performance gains are from the refactored adjacency sort.)

I determined that the bug I was hunting is because adjacency is now functionally defined as any face that has >2 previously observed vertices. This allows a face (triangle 4 in the photo) to be added, even if all of its edges are new/previously unobserved. The orientation of this triangle cannot be checked against existing faces, and we can end up with irreparable inconsistencies when the "hole" (triangle 5 in the photo) is filled (if the orientation of triangle 4 is ultimately inconsistent with the correct orientation of triangle 5).
<img src="https://github.com/user-attachments/assets/f343b6bb-7e6b-438e-8695-8ae24d8f5f8d" width="436" height="500" />

That bug is not present in the original code on `master`.